refactor: move the allocation functions into a new allocation.c file
Pure move of _ALLOCATE_AVAILABLE and its normal-distribution helpers,
plus _ALLOCATE_BY_PRIORITY and its __sum helper (added in #809), out
of vensim.c in preparation for additional allocation functions. Also
remove redundant extern declarations that are already provided by
sde.h.

One non-move change: the unconditional
'#define PRINT_ALLOCATIONS_DEBUG_INFO' that came in with #809 is
commented out. It appears to be vestigial (nothing between it and the
end of vensim.c used it), and leaving it active would have enabled
the debug printing in the allocation functions that now follow it in
allocation.c.
