+++ packages/cli/src/c/macros.c
@@ -1,4 +1,1 @@
 #include "sde.h"
-
-extern double _time;
-extern double _time_step;
